day 5 part 1 commit

+++ aoc-2023/Helpers.cs
@@ -23,7 +23,7 @@ namespace aoc_2023
         /// <summary>
         /// <para>Outputs the answer to the console.</para>
         /// </summary>
-        public static void OutputSolve(int day, int part, float solution)
+        public static void OutputSolve(int day, int part, double solution)
         {
             Console.ForegroundColor = ConsoleColor.DarkGreen;
             Console.Write($"Day ");
